+++ Part 2/csharp_solution/Exchange/Trade.cs
@@ -4,7 +4,7 @@ namespace Exchange
 {
     public struct Trade
     {
-        public Trade(string buyer, string seller, string instrument, long quantity, decimal price)
+        public Trade(string buyer, string seller, string instrument, long quantity, double price)
         {
             Buyer = buyer;
             Seller = seller;
@@ -17,6 +17,6 @@ namespace Exchange
         public string Seller { get; }
         public string Instrument { get; }
         public long Quantity { get; }
-        public decimal Price { get; }
+        public double Price { get; }
     }
 }
